ui/settings: isolate handlers to its own modules
