+++ src/libtrx/game/ui/dialogs/setting_tabs/sound_settings.def
@@ -1,5 +1,5 @@
-X_UI_CFG_FLOAT_PERCENT(audio.sound_volume,              SOUND_SETTINGS_SOUND_VOLUME, .custom_handler = { .request_change_value = M_Volume_RequestChange }, .min_value = 0, .max_value = 100, .delta_slow = 10, .delta_fast = 10)
-X_UI_CFG_FLOAT_PERCENT(audio.music_volume,              SOUND_SETTINGS_MUSIC_VOLUME, .custom_handler = { .request_change_value = M_Volume_RequestChange }, .min_value = 0, .max_value = 100, .delta_slow = 10, .delta_fast = 10)
+X_UI_CFG_FLOAT_PERCENT(audio.sound_volume,              SOUND_SETTINGS_SOUND_VOLUME, .custom_handler = { .request_change_value = UI_Settings_Volume_RequestChange }, .min_value = 0, .max_value = 100, .delta_slow = 10, .delta_fast = 10)
+X_UI_CFG_FLOAT_PERCENT(audio.music_volume,              SOUND_SETTINGS_MUSIC_VOLUME, .custom_handler = { .request_change_value = UI_Settings_Volume_RequestChange }, .min_value = 0, .max_value = 100, .delta_slow = 10, .delta_fast = 10)
 
 #if TR_VERSION == 1
 X_UI_CFG_BOOL(audio.enable_music_in_menu,               SOUND_SETTINGS_ENABLE_MUSIC_IN_MENU)
@@ -8,8 +8,8 @@ X_UI_CFG_BOOL(audio.fix_speeches_killing_music,         SOUND_SETTINGS_FIX_SPEEC
 #endif
 
 X_UI_CFG_BOOL_INV(audio.enable_music_in_inventory,      SOUND_SETTINGS_PAUSE_MUSIC_IN_INVENTORY)
-X_UI_CFG_FLOAT_PERCENT(audio.inventory_ambient_volume,  SOUND_SETTINGS_INVENTORY_AMBIENT_VOLUME, .min_value = 0, .max_value = 100, .delta_slow = 10, .delta_fast = 10, .custom_handler = { .is_available = M_PauseMusicInInventory_IsAvailable})
-X_UI_CFG_FLOAT_PERCENT(audio.inventory_music_volume,    SOUND_SETTINGS_INVENTORY_MUSIC_VOLUME, .min_value = 0, .max_value = 100, .delta_slow = 10, .delta_fast = 10, .custom_handler = { .is_available = M_PauseMusicInInventory_IsAvailable })
+X_UI_CFG_FLOAT_PERCENT(audio.inventory_ambient_volume,  SOUND_SETTINGS_INVENTORY_AMBIENT_VOLUME, .min_value = 0, .max_value = 100, .delta_slow = 10, .delta_fast = 10, .custom_handler = { .is_available = UI_Settings_PauseMusicInInventory_IsAvailable})
+X_UI_CFG_FLOAT_PERCENT(audio.inventory_music_volume,    SOUND_SETTINGS_INVENTORY_MUSIC_VOLUME, .min_value = 0, .max_value = 100, .delta_slow = 10, .delta_fast = 10, .custom_handler = { .is_available = UI_Settings_PauseMusicInInventory_IsAvailable })
 X_UI_CFG_FLOAT_PERCENT(audio.underwater_ambient_volume, SOUND_SETTINGS_UNDERWATER_AMBIENT_VOLUME, .min_value = 0, .max_value = 100, .delta_slow = 10, .delta_fast = 10)
 X_UI_CFG_FLOAT_PERCENT(audio.underwater_music_volume,   SOUND_SETTINGS_UNDERWATER_MUSIC_VOLUME, .min_value = 0, .max_value = 100, .delta_slow = 10, .delta_fast = 10)
 X_UI_CFG_ENUM(audio.music_load_condition,               SOUND_SETTINGS_MUSIC_LOAD_CONDITION, .misc = UI_Settings_MusicLoadConditionEnumEntries)
