+++ src/libtrx/meson.build
@@ -253,6 +253,9 @@ sources = [
   'game/ui/dialogs/save_slot.c',
   'game/ui/dialogs/select_level.c',
   'game/ui/dialogs/setting_helpers/enums.c',
+  'game/ui/dialogs/setting_helpers/handlers.c',
+  'game/ui/dialogs/setting_helpers/handlers_language.c',
+  'game/ui/dialogs/setting_helpers/handlers_screen.c',
   'game/ui/dialogs/settings.c',
   'game/ui/dialogs/sound_settings.c',
   'game/ui/dialogs/stats.c',
